updated app structure
