+++ lib/config/assets.dart
@@ -1,0 +1,3 @@
+abstract final class Assets {
+  static const _basePath = "assets/images";
+}
