+++ lib/core/services/locale_service.dart
@@ -1,0 +1,35 @@
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:flutter_template/common_libs.dart';
+import 'package:flutter_template/core/services/app_service.dart';
+import 'package:intl/intl_standalone.dart';
+
+class LocaleService {
+  AppLocalizations? _strings;
+  AppLocalizations get strings => _strings!;
+
+  bool get isLoaded => _strings != null;
+
+  bool get isEnglish => strings.localeName == 'en';
+
+  Future<void> load() async {
+    final localeCode =
+        settingsLogic.currentLocale.value ?? await findSystemLocale();
+    Locale locale = Locale(localeCode.split('_')[0]);
+    if (kDebugMode) {
+      // Uncomment for testing in chinese
+      // locale = Locale('zh');
+    }
+    if (AppLocalizations.supportedLocales.contains(locale) == false) {
+      locale = const Locale('en');
+    }
+    settingsLogic.currentLocale.value = locale.languageCode;
+    _strings = await AppLocalizations.delegate.load(locale);
+  }
+
+  Future<void> loadIfChanged(Locale locale) async {
+    bool didChange = _strings?.localeName != locale.languageCode;
+    if (didChange && AppLocalizations.supportedLocales.contains(locale)) {
+      _strings = await AppLocalizations.delegate.load(locale);
+    }
+  }
+}
