+++ lib/ui/core/themes/colors.dart
@@ -1,0 +1,50 @@
+import 'package:flutter_template/common_libs.dart';
+
+abstract final class AppColors {
+  static const black1 = Color(0xFF101010);
+  static const white1 = Color(0xFFFFF7FA);
+  static const grey1 = Color(0xFFF2F2F2);
+  static const grey2 = Color(0xFF4D4D4D);
+  static const grey3 = Color(0xFFA4A4A4);
+  static const whiteTransparent = Color(0x4DFFFFFF);
+  static const blackTransparent = Color(0x4D000000);
+  static const red1 = Color(0xFFE74C3C);
+  
+  /// Common
+  static const white = Colors.white;
+  static const black = Colors.black;
+  static const midnightGreen = Color(0xFF00555E);
+  static const silverPink =  Color(0xFFBEABA1);
+  static const offWhite =  Color(0xFFF8ECE5);
+  static const error = Colors.red;
+
+  static const lightColorScheme = ColorScheme(
+    brightness: Brightness.light,
+    primary: AppColors.black1,
+    onPrimary: AppColors.white1,
+    secondary: AppColors.black1,
+    onSecondary: AppColors.white1,
+    surface: Colors.white,
+    onSurface: AppColors.black1,
+    error: Colors.white,
+    onError: Colors.red,
+  );
+
+  static const darkColorScheme = ColorScheme(
+    brightness: Brightness.dark,
+    primary: AppColors.white1,
+    onPrimary: AppColors.black1,
+    secondary: AppColors.white1,
+    onSecondary: AppColors.black1,
+    surface: AppColors.black1,
+    onSurface: Colors.white,
+    error: Colors.black,
+    onError: AppColors.red1,
+  );
+
+  static const greyFade = const LinearGradient(
+    begin: Alignment.topCenter,
+    end: Alignment.bottomCenter,
+    colors: [Color(0xFF626161), Color(0xFF212121)],
+  );
+}
